Enable drm-kms-egl on bookworm and quiet drm-cxx/shell warnings
build_pi.sh: add --with-local-display-info, which cross-builds a static
libdisplay-info (>=0.2.0, pinned by tag) into the sysroot via meson so the
drm-cxx backend can be built against distros that ship an older one
(bookworm has 0.1.1). Linked statically, so no new runtime dependency on
the target. When the flag is set, the apt libdisplay-info-dev is skipped
and its dev symlink removed so the static archive wins at link time, and
drm-kms-egl is no longer filtered out of bookworm.

build_pi.sh: move libwayland-dev and wayland-protocols into the shared
sysroot deps. waypp is always built and unconditionally requires
wayland-client/cursor/protocols, so a drm-kms-egl- or software-only build
needs them too.

shell/CMakeLists.txt: force EGL/GLESv2 as DT_NEEDED for drm-kms-egl. They
sit ahead of platform_homescreen (which references glDeleteTextures) in the
link line, so --as-needed could drop libGLESv2 before the reference is seen.

drm_kms.cmake: treat the vendored drm-cxx as a system library (SYSTEM
interface includes + silence its own-source warnings); third_party
CMakeLists silences waypp's own-source warnings the same way.

shell: fix sign-conversion and missing-field-initializer warnings in
drm_backend, driver_probe, drm_seat, and flutter_view (static_casts and
explicit field initializers); clang-format the touched files.

+++ shell/backend/drm_kms_egl/driver_probe.cc
@@ -61,7 +61,8 @@ std::string GetDriverName(const int drm_fd) {
   if (!v) {
     return {};
   }
-  std::string name(v->name ? v->name : "", v->name_len > 0 ? v->name_len : 0);
+  std::string name(v->name ? v->name : "",
+                   static_cast<size_t>(v->name_len > 0 ? v->name_len : 0));
   drmFreeVersion(v);
   return name;
 }
